// web/src/api/generated/models/preventiveMaintenanceAcknowledgementResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * UniPM.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PreventiveMaintenanceAcknowledgementResponse {
  id: string
  formId: string
  signatoryName: string
  signatoryPosition: string
  signatureContentType: string
  signatureChecksum: string
  capturedByUserId: string
  acknowledgedAt: string
}